[issue693] let hash function return unsigned int instead of size_t

+++ src/search/state_registry.h
@@ -111,7 +111,7 @@ class StateRegistry {
               state_size(state_size) {
         }
 
-        size_t operator()(StateID id) const {
+        unsigned int operator()(StateID id) const {
             return utils::hash_unsigned_int_sequence(
                 state_data_pool[id.value], state_size);
         }
